requirements.txt
tensorflow>=2.12.0
tensorflow-hub>=0.13.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
